A LOT of cleanup and refactoring of NavBar
Now, NavBar only needs to know about what page it's on. Also, every page
defines a method called getMainWindow() which returns the window which,
for the purposes of the nav bar, is the main window of that page. This
is the window whose status/edit pages will get shown when you click the
buttons on the nav bar from a page.

For example: I'm on the status page for a window. The status page knows
about which Window object it's a status page for, so it tells its NavBar
instance that when the edit button is clicked, it should show the edit
page for the current status page's window.

The main page and presets page just tell their NavBars to go to window 1
when the edit or status buttons are clicked. EVERY page has a different
instance of a NavBar, including things like CameraRoll. and EVERY page
defines a getMainWindow() method to tell the NavBar which page to go to.

ALSO: there is ONLY ONE PLACE where the windows are defined, and it's in
the app main.js.

Alas, this is a much better way, architecturally, for NavBar to work,
but it doesn't fix any of the bugs that are happening.